archive/requirements.txt
Flask>=2.2.0
Pillow>=6.1.0